Move Transform and Formatter into their own files
